requirements.txt
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==2.2
django-allauth==0.41.0
django-braces==1.14.0
django-cors-headers==3.2.1
django-cors-middleware==1.5.0
django-oauth-toolkit==1.2.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-rest-framework-social-oauth2==1.1.0
django-rest-passwordreset==1.0.0
djangorestframework==3.11.0
fcm-django==0.3.4
idna==2.8
oauthlib==3.1.0
Pillow==7.0.0
pyfcm==1.4.7
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
urllib3==1.25.7
gunicorn==20.1.0